requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.1
Django==2.2.6
django-filter==2.2.0
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.10.3
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==1.20.0
idna==2.8
Markdown==3.1.1
mongoengine==0.18.2
numpy==1.17.3
pandas==0.25.3
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
ratelim==0.1.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
xlrd==1.2.0
